refactor(db): extract inline tests from 3 db modules (partial #413)
Move the inline #[cfg(test)] mod tests blocks from auth/login.rs,
auth/device.rs, and ebook/accent.rs into sibling tests.rs files,
following the db/src/books/tests.rs pattern called out in
.claude/rules/03-unit-testing.md.

Pure mechanical extraction: identical test bodies, identical use
statements, replaced inline blocks with `#[cfg(test)] mod tests;`.
All 19 affected tests still pass. No production code changed.

audiobook.rs (listed in the issue) already has a sibling tests.rs;
auth/device.rs ships here as a fourth file alongside the originally
prioritized three.

Refs #413
